+++ package.json
@@ -27,6 +27,7 @@
     "express": "^4.17.1",
     "moment": "^2.24.0",
     "p-event": "^4.1.0",
+    "p-timeout": "^3.2.0",
     "urijs": "^1.19.2"
   },
   "devDependencies": {
