+++ src/util/sleep.ts
@@ -1,0 +1,3 @@
+export async function sleep(ms: number): Promise<void> {
+  return new Promise(resolve => setTimeout(resolve, ms));
+}
